src/app/multi-task-question-answering-assistant/requirements.txt
# LangChain 1.x 核心依赖
langchain>=1.0.0
langchain-core>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# OpenAI API
openai>=1.12.0

# 数据验证
pydantic>=2.7.4
pydantic-settings>=2.3.4

# 环境变量管理
python-dotenv>=1.0.0

# 日志
loguru>=0.7.2

# Web 框架
fastapi>=0.104.1
uvicorn>=0.24.0

# HTTP 客户端
httpx>=0.25.0
requests>=2.31.0

# 搜索 API
tavily-python>=0.3.3

# Redis（可选）
redis>=5.0.1

# 开发工具
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=24.1.0
ruff>=0.1.0
